Fix crash after failing to load SP game library
When loading a game in SP, UI_FreeAllSpecies can be called twice without an initialization in between if the SP game library could not be loaded. This results in a double free crash. This commit fixes the issue by resetting some variables after freeing so that calling free again is a no-op.